+++ src/anyvar/storage/__init__.py
@@ -1,5 +1,5 @@
 """Provide tools for and implementations of AnyVar storage backends."""
 
-from .abc import _Storage
+from .base_storage import Storage
 
 DEFAULT_STORAGE_URI = "postgresql://postgres@localhost:5432/anyvar"
