+++ src/anyvar/storage/postgres.py
@@ -7,7 +7,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.dialects.postgresql import insert
 from sqlalchemy.orm import joinedload, sessionmaker
 
-from anyvar.storage.abc import Storage, StoredObjectType, VariationMappingType
+from anyvar.storage.base_storage import Storage, StoredObjectType, VariationMappingType
 from anyvar.storage.mapper_registry import mapper_registry
 from anyvar.storage.orm import (
     Allele,
