+++ tests/test_lifespan.py
@@ -2,13 +2,13 @@ from fastapi import FastAPI
 from fastapi.testclient import TestClient
 
 from anyvar.restapi.main import app_lifespan
-from anyvar.storage import _Storage
+from anyvar.storage import Storage
 
 
 def test_lifespan(mocker):
     """Test the app_lifespan method in anyvar.restapi.main"""
     create_storage_mock = mocker.patch("anyvar.anyvar.create_storage")
-    storage_mock = mocker.Mock(spec=_Storage)
+    storage_mock = mocker.Mock(spec=Storage)
     create_storage_mock.return_value = storage_mock
     create_translator_mock = mocker.patch("anyvar.anyvar.create_translator")
     create_translator_mock.return_value = {}
